+++ pysyncgateway/__init__.py
@@ -4,6 +4,7 @@ from .admin_client import AdminClient
 from .database import Database
 from .document import Document
 from .query import Query
+from .session import Session
 from .stats import Stats
 from .user import User
 from .user_client import UserClient
@@ -13,6 +14,7 @@ __all_ = [
     Database,
     Document,
     Query,
+    Session,
     Stats,
     User,
     UserClient,